refactor : idx 데이터 타입 long으로 변경

+++ src/main/java/shop/makaroni/bunjang/src/dao/UserDao.java
@@ -122,7 +122,7 @@ public class UserDao {
 		return itemMapper.searchStoreItemByName(userId, itemName, condition, pagingCond);
 	}
 
-	public int checkUserIdx(int userIdx){
+	public int checkUserIdx(long userIdx){
 		String query = "select exists(select idx from User where idx = ?)";
 		return this.jdbcTemplate.queryForObject(query,
 				int.class,
